[DO NOT MERGE] NLA: Go back to DDV_sprout Adreno and SDM blobs
Somehow StatiX doesn't play well with UM9 Adreno blobs, resulting
in splash stuck forever even though SurfaceFlinger told that it's
booted. Also uses in-tree device specific HAL, it's kinda weird for
our device at this state right now, but at this rate let's just use
it.

Also pin the blobs in order to not getting overwritten next time
if we want to

TLDR: This is an experimental commit, please do not merge.

Signed-off-by: Tuan Anh <tuan73176@gmail.com>