NLA: Hexpatch libfp_client for new Parcel size
Parcel object increased from 104 to 120 bytes, and many functions in
both libraries had two Parcel objects next to each other in their stack.

Each of those functions had their stack modified to be increased by
32 bytes to accommodate for the increased Parcel object size. The stack
pointer offsets for the 2nd Parcel, along with any other items in the
stack after the Parcels (which was just the stack canary), were also
adjusted to accomodate for the change in Parcel size.